Replace syntax functions
