+++ benchmark/src/main/scala/benchmark/wrapper/ldbc/Insert.scala
@@ -23,8 +23,7 @@ import ldbc.sql.*
 
 import ldbc.dsl.SQL
 
-import ldbc.query.builder.syntax.*
-import ldbc.query.builder.Table
+import ldbc.query.builder.*
 
 import jdbc.connector.*
