+++ module/ldbc-query-builder/src/test/scala/ldbc/query/builder/QuerySyntaxTest.scala
@@ -13,8 +13,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
   it should "It can be converted to the model specified by the toList method." in {
     assertCompiles(
       """
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -28,8 +27,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
   it should "If the Tuple generated by the toList method does not match the type of the specified model property, a Compile error occurs." in {
     assertDoesNotCompile(
       """
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -46,8 +44,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -61,8 +58,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
   it should "It can be converted to the model specified by the headOption method." in {
     assertCompiles(
       """
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -76,8 +72,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
   it should "If the Tuple generated by the headOption method does not match the type of the specified model property, a Compile error occurs." in {
     assertDoesNotCompile(
       """
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -94,8 +89,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -111,8 +105,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -126,8 +119,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
   it should "If the Tuple generated by the unsafe method does not match the type of the specified model property, a Compile error occurs." in {
     assertDoesNotCompile(
       """
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -144,8 +136,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int) derives Table
         |
@@ -161,8 +152,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int, categoryId: Long) derives Table
         |
@@ -186,8 +176,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
       """
         |import cats.effect.IO
         |
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class User(id: Long, name: String, age: Int, categoryId: Long) derives Table
         |
@@ -212,8 +201,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
         |import cats.effect.IO
         |
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class Country(code: String, name: String) derives Table
         |
@@ -243,8 +231,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
         |import cats.effect.IO
         |
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class Country(code: String, name: String) derives Table
         |
@@ -274,8 +261,7 @@ class QuerySyntaxTest extends AnyFlatSpec:
         |import cats.effect.IO
         |
         |import ldbc.dsl.DBIO
-        |import ldbc.query.builder.Table
-        |import ldbc.query.builder.syntax.*
+        |import ldbc.query.builder.*
         |
         |case class Country(code: String, name: String) derives Table
         |
