+++ module/ldbc-statement/src/main/scala/ldbc/statement/Command.scala
@@ -8,12 +8,46 @@ package ldbc.statement
 
 import scala.annotation.targetName
 
-import ldbc.dsl.{ Parameter, SQL }
+import ldbc.dsl.{ Parameter, SQL, DBIO }
+import ldbc.dsl.codec.Decoder
 
 /**
  * Trait represents a command in MySQL.
  */
-trait Command extends SQL
+trait Command extends SQL:
+
+  /**
+   * A method to execute an update operation against the MySQL server.
+   *
+   * {{{
+   *   TableQuery[User]
+   *     .update(user => user.id *: user.name *: user.age)((1L, "Alice", 20))
+   *     .where(_.id === 1L)
+   *     .update
+   * }}}
+   *
+   * @return
+   * The number of rows updated
+   */
+  def update: DBIO[Int] = DBIO.update(statement, params)
+
+  /**
+   * A method to execute an insert operation against the MySQL server.
+   *
+   * {{{
+   *   TableQuery[User]
+   *     .insertInto(user => user.name *: user.age)
+   *     .values(("Alice", 20))
+   *     .returning[Long]
+   * }}}
+   *
+   * @tparam T
+   * The type of the primary key
+   * @return
+   * The primary key value
+   */
+  def returning[T <: String | Int | Long](using decoder: Decoder[T]): DBIO[T] =
+    DBIO.returning(statement, params, decoder)
 
 object Command:
   case class Pure(statement: String, params: List[Parameter.Dynamic]) extends Command:
